reversing question container
